Revised rewards
Reward did not show up on the rewards list if you did not go to the page before visiting another giant. This is now fixed so you do not have to see the reward. You can now visit them all without displaying the reward first and still be able to see them.